de verschillende processen met enkele recursie weer een beetje uit elkaar gehaald

+++ scripts/helper/rename.py
@@ -1,44 +1,50 @@
-from venv.flac.services import(
+from venv.flac.services import (
     has_haakjes, replace_haakjes,)
 import os
-import glob
 
 cover_names = ['box front', 'front', 'Cover', 'cover']
 cover_nice = 'folder.jpg'
 
 
-def rename_cover(path):
-    for name in cover_names:
-        src = '{}/{}.jpg'.format(path, name)
+def rename_cover_one(path, name):
+    src = '{}/{}.jpg'.format(path, name)
+    if os.path.exists(src):
         trg = '{}/{}'.format(path, cover_nice)
-        if os.path.exists(trg):
-            return
-        if os.path.exists(src):
+        if not os.path.exists(trg):
             os.rename(src, trg)
             print('renamed to:{}'.format(trg))
 
 
-def sanatize_haakjes(path, d):
-    # print(d)
-    if has_haakjes(d):
-        src = '{}/{}'.format(path, d)
-        dst = '{}/{}'.format(path, replace_haakjes(d))
-        if os.path.exists(src):
+def rename_cover(path, step_in):
+    for name in cover_names:
+        rename_cover_one(path, name)
+        if step_in:
+            # one recursive step
+            for d2 in os.listdir(path):
+                p2 = u'{}/{}'.format(path, d2)
+                if os.path.isdir(p2):
+                    rename_cover_one(p2, d2)
+
+
+def sanatize_haakjes_one(path, d):
+    src = u'{}/{}'.format(path, d)
+    if os.path.exists(src) and os.path.isdir(src):
+        if has_haakjes(d):
+            d_trg = replace_haakjes(d)
+            dst = u'{}/{}'.format(path, d_trg)
             os.rename(src, dst)
             print(dst)
+            return dst
+        else:
+            return src
+    return None
 
 
-# def process_path(path, f):
-#     print(f)
-#     p = '{}/{}'.format(path, f)
-#     if os.path.isdir(p):
-#         for ff in os.listdir(p):
-#             print(ff)
-#             src = u'{}/{}'.format(p, ff)
-#             trg = u'{}/{}'.format(os.path.dirname(p), ff)
-#             print(src)
-#             print(trg)
-#             print('--')
-#             # os.rename(p, trg)
-
-
+def sanatize_haakjes(path, step_in):
+    for d in os.listdir(path):
+        dst = sanatize_haakjes_one(path, d)
+        if dst and step_in:
+            # one recursive step
+            for d2 in os.listdir(dst):
+                p2 = u'{}/{}'.format(dst, d2)
+                sanatize_haakjes_one(p2, d2)
